+++ include/compute/compute_partial_derivatives.hpp
@@ -3,8 +3,6 @@
 
 #include "kokkos_abstractions.h"
 #include "quadrature.h"
-#include "receiver.h"
-#include "source.h"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
 #include <vector>
