+++ include/compute/compute_receivers.hpp
@@ -3,8 +3,7 @@
 
 #include "kokkos_abstractions.h"
 #include "quadrature.h"
-#include "receiver.h"
-#include "source.h"
+#include "receiver/interface.hpp"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
 #include <vector>
