+++ include/receiver/interface.hpp
@@ -1,0 +1,7 @@
+#ifndef _RECEIVER_INTERFACE_HPP
+#define _RECEIVER_INTERFACE_HPP
+
+#include "read_receiver.hpp"
+#include "receiver.hpp"
+
+#endif
