+++ include/source/interface.hpp
@@ -1,0 +1,9 @@
+#ifndef _SOURCE_INTERFACE_HPP
+#define _SOURCE_INTERFACE_HPP
+
+#include "force_source.hpp"
+#include "moment_tensor_source.hpp"
+#include "read_sources.hpp"
+#include "source.hpp"
+
+#endif
