+++ src/compute/compute_sources.cpp
@@ -1,7 +1,7 @@
 #include "compute/interface.hpp"
 #include "kokkos_abstractions.h"
 #include "quadrature.h"
-#include "source.h"
+#include "source/interface.hpp"
 #include "specfem_mpi.h"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
