+++ src/source/read_sources.cpp
@@ -1,0 +1,44 @@
+#include "source/interface.hpp"
+
+std::tuple<std::vector<specfem::sources::source *>, type_real>
+specfem::sources::read_sources(const std::string sources_file,
+                               const type_real dt,
+                               const specfem::MPI::MPI *mpi) {
+  // read sources file
+  std::vector<specfem::sources::source *> sources;
+  YAML::Node yaml = YAML::LoadFile(sources_file);
+  int nsources = yaml["number-of-sources"].as<int>();
+  YAML::Node Node = yaml["sources"];
+  assert(Node.IsSequence());
+  for (auto N : Node) {
+    if (YAML::Node force_source = N["force"]) {
+      sources.push_back(new specfem::sources::force(force_source, dt));
+    } else if (YAML::Node moment_tensor = N["moment-tensor"]) {
+      sources.push_back(new specfem::sources::moment_tensor(moment_tensor, dt));
+    }
+  }
+
+  if (sources.size() != nsources) {
+    std::ostringstream message;
+    message << "Found only " << sources.size()
+            << " number of sources. Total number of sources in " << sources_file
+            << " are" << nsources
+            << " Please check if there is a error in sources file.";
+    throw std::runtime_error(message.str());
+  }
+
+  type_real t0 = std::numeric_limits<type_real>::max();
+  for (auto &source : sources) {
+    type_real cur_t0 = source->get_t0();
+    if (cur_t0 < t0) {
+      t0 = cur_t0;
+    }
+  }
+
+  for (auto &source : sources) {
+    type_real cur_t0 = source->get_t0();
+    source->update_tshift(cur_t0 - t0);
+  }
+
+  return std::make_tuple(sources, t0);
+}
